[Refactor] vm의 UIkit 관련 기능들 제거(텍스트 width, height 구하는 로직 등)
