+++ travelPlan/Source/Feature/Feed/ViewModel/FeedViewModel.swift
@@ -5,7 +5,7 @@
 //  Created by 양승현 on 2023/05/12.
 //
 
-import UIKit
+import Foundation
 import Combine
 
 // notificationArrivedTODO: - NotificationIcon render
