+++ travelPlan/Source/Feature/Feed/ViewModel/Post/PostCellViewModel.swift
@@ -5,7 +5,7 @@
 //  Created by 양승현 on 2023/05/14.
 //
 
-import UIKit
+import Foundation
 
 final class PostCellViewModel {
   // MARK: - Properties
